// libs/ui/src/lib/button/button.enum.ts
export enum ButtonTypeEnum {
  BUTTON = 'button',
  TEXT = 'text',
}

export enum ButtonColorTypeEnum {
  DANGER = 'danger',
  PRIMARY = 'primary',
  PRIMARY_OUTLINE = 'primary_outline',
  // WARNING = 'warning',
  SECONDARY = 'secondary',
  SECONDARY_OUTLINE = 'secondary_outline',
  WHITE = 'white',
  ICON_CIRCLE = 'icon_circle',
}

export enum ButtonFillTypeEnum {
  SOLID = 'solid',
  OUTLINE = 'outline',
  CLEAR = 'clear',
  CLEAR_NO_UPPERCASE = 'clear_no_uppercase',
}

export enum ButtonExpandTypeEnum {
  BLOCK = 'block',
  FULL = 'full',
}

// export enum ButtonSizeTypeEnum {
//   DEFAULT = 'default',
//   SMALL = 'small',
//   LARGE = 'large',
// }

export enum ButtonShapeTypeEnum {
  ROUND = 'round',
}